Initial Setup: Data Structures & Algorithms.
The idea behind is to review concepts of Data Structures
and algorithms. It's baby steps like reviewing so for many
developers would find this boring.
It's a collection from tutorials/talks I've found useful.

I will be adding the code on ruby, js and python .. or any languange
I'm playing / learning with in my playground